Fix compile warnings (unused-but-set-variable) in groupwise protocol

+++ protocols/groupwise/libgroupwise/inputprotocolbase.cpp
@@ -98,7 +98,7 @@ bool InputProtocolBase::safeReadBytes( QByteArray & data, uint & len )
 		// so look for that in the last position instead of \0
 		// this caused a crash - guessing that temp.length() is set to the number of bytes actually read...
 		// if ( (quint8)( * ( temp.data() + ( temp.length() - 1 ) ) ) == 0xFF )
-		if ( temp.length() < ( val - 1 ) )
+		if ( (quint32)temp.length() < ( val - 1 ) )
 		{
 			debug( QString( "InputProtocol::safeReadBytes() - string broke, giving up, only got: %1 bytes out of %2" ).arg( temp.length() ).arg( val ) );
 			m_state = NeedMore;
